requirements.txt
MacFSEvents==0.4
colorama==0.3.3
freezegun==0.3.1
mock==1.0.1
mockredispy==2.9.0.10
nose==1.3.4
python-dateutil==2.4.0
python-termstyle==0.1.10
redis==2.10.3
six==1.9.0
sniffer==0.3.4
wsgiref==0.1.2
